04_mongo/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
pymongo


# pandas
# redis

# prometheus_client
# requests
# pika
# confluent-kafka
